n8n-management-app/requirements.txt
# Core dependencies for n8n Management App
docker>=6.0.0
pyyaml>=6.0
requests>=2.28.0
psutil>=5.9.0
tkinter-tooltip>=2.0.0